+++ x/tokenomics/types/codec.go
@@ -8,5 +8,6 @@ import (
 func RegisterInterfaces(registrar codectypes.InterfaceRegistry) {
 	registrar.RegisterImplementations((*sdk.Msg)(nil),
 		&MsgUpdateParams{},
+		&MsgUpdateSlashingDowntimeParams{},
 	)
 }